# lab2/payment-auth-system/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.CreateKeyRequest:
    properties:
      key_value:
        type: string
    type: object
  handlers.CreateTerminalRequest:
    properties:
      address:
        type: string
      name:
        type: string
      serial_number:
        type: string
      status:
        type: boolean
    type: object
...
  handlers.UpdateKeyRequest:
    properties:
      key_value:
        type: string
    type: object
  handlers.UpdateTerminalRequest:
    properties:
      address:
        type: string
      name:
        type: string
      serial_number:
        type: string
      status:
        type: boolean
    type: object
...
  models.Key:
    properties:
      createdAt:
        type: string
      id:
        format: int64
        type: integer
      keyValue:
        type: string
    type: object
  models.Terminal:
    properties:
      address:
        type: string
      createdAt:
        type: string
      id:
        format: int64
        type: integer
      name:
        type: string
      serialNumber:
        type: string
      status:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: REST API сервера авторизации платежей транспортными картами
  title: Payment Auth System API
  version: "1.0"
paths:
...
  /keys:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Key'
            type: array
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Получить список ключей
      tags:
      - keys
    post:
      consumes:
      - application/json
      parameters:
      - description: Create key request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateKeyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Создать ключ
      tags:
      - keys
  /keys/{id}:
    delete:
      parameters:
      - description: Key ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Удалить ключ
      tags:
      - keys
    get:
      parameters:
      - description: Key ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Key'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Получить ключ по ID
      tags:
      - keys
    put:
      consumes:
      - application/json
      parameters:
      - description: Key ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update key request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateKeyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Обновить ключ
      tags:
      - keys
...
  /terminals:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Terminal'
            type: array
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Получить список терминалов
      tags:
      - terminals
    post:
      consumes:
      - application/json
      parameters:
      - description: Create terminal request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateTerminalRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Создать терминал
      tags:
      - terminals
  /terminals/{id}:
    delete:
      parameters:
      - description: Terminal ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Удалить терминал
      tags:
      - terminals
    get:
      parameters:
      - description: Terminal ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Terminal'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Получить терминал по ID
      tags:
      - terminals
    put:
      consumes:
      - application/json
      parameters:
      - description: Terminal ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update terminal request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateTerminalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Обновить терминал
      tags:
      - terminals
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"